+++ CHANGELOG.md
@@ -4,3 +4,6 @@ v3:
 * Added support for `version` parameter, to handle `/v2` URLs
 * Retry 500 errors
 * Add documentation for working with IAPs
+
+v4:
+* Added support for `read()` and `readAll()`, a clearer API than `fetchJson` and its `options` object